+++ src/main/java/org/moera/lib/node/types/PublicMediaFileInfo.java
@@ -12,8 +12,8 @@ public class PublicMediaFileInfo implements Cloneable {
     private String path;
     private Integer width;
     private Integer height;
-    private Integer orientation;
-    private int size;
+    private Short orientation;
+    private long size;
 
     @JsonIgnore
     private Object extra;
@@ -50,19 +50,19 @@ public class PublicMediaFileInfo implements Cloneable {
         this.height = height;
     }
 
-    public Integer getOrientation() {
+    public Short getOrientation() {
         return orientation;
     }
 
-    public void setOrientation(Integer orientation) {
+    public void setOrientation(Short orientation) {
         this.orientation = orientation;
     }
 
-    public int getSize() {
+    public long getSize() {
         return size;
     }
 
-    public void setSize(int size) {
+    public void setSize(long size) {
         this.size = size;
     }
